+++ neo4j/_async_compat/util.py
@@ -21,6 +21,12 @@ import inspect
 from ..meta import experimental
 
 
+__all__ = [
+    "AsyncUtil",
+    "Util",
+]
+
+
 class AsyncUtil:
     @staticmethod
     async def iter(it):